% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2wkt.R
\name{df2wkt}
\alias{df2wkt}
\title{Writes a data.frame as 'well-known-text' or 'wkt'.}
\usage{
df2wkt(df, fn = "tmp_wkt.txt")
}
\arguments{
\item{df}{Stratum boundary for a single stratum as a tibble/data.frame with cols 'lat' and 'lon'}

\item{fn}{Filename to store wkt representation in.}
}
\value{
A wkt-text-file written out.
}
\description{
Takes a data.frame/tibble with coordinates / stratum boundary and
writes them out as 'well-know-text' or wkt-file. Assumes open polygon
i.e. first line in df is appended before writing out, as the result is
to be used as input to 'Rstox::surveyPlanner' or as step in
the wrapper 'hafrostox::srv_plnr'.
}
\examples{
\dontrun{

tibble::tribble(~name, ~lat, ~lon,
  "NW1",  63.56, -20.57, 	  	  #1
  "NE", 63.44, -20.11,		  #2
  "SE", 63.31, -20.27,		  #3
  "SW", 63.28, -20.65,		  #4
  "NW2", 63.49,-20.64) -> stratum 	  #5

df2wkt(stratum, "stratum_wkt.txt")
}
}
\keyword{parameters}
\keyword{stratum,}
\keyword{surveyPlanner,}
