% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_df_trans.R
\name{flip_df_trans}
\alias{flip_df_trans}
\title{Flips a transect layout given as a dataframe.}
\usage{
flip_df_trans(tr_df)
}
\arguments{
\item{tr_df}{Dataframe containing a transect layout.}
}
\value{
An flipped transect layout as a dataframe.
}
\description{
Transect layout flipped by changing the order of the nodes.
}
\details{
Note that number of nodes must be even for the flip to work.
}
\examples{
\dontrun{
  library(geo)
  tibble::tribble(~lat,~lon,
    67, -23,
    68, -23, 68, -22,
    67, -22, 67, -21,
    68, -21, 68, -20,
    67, -20, 67, -19,
    68, -19, 68, -18,
    67, -18, 67, -17,
    68, -17, 68, -16,
    67, -16, 67, -15,
    68, -15, 68, -14,
    67, -14) -> df
 geo::geoplot(df,type="n",grid=F,xlim=list(lat=c(66,68.25),lon=c(-24,-13)))
 geo::gbplot(c(500,1000),lty=2)
 geo::geolines(df,col="blue",lwd=5,lty=2)
 geo::geolines(flip_df_trans(df),col="magenta",lwd=3)
}
}
\keyword{transect}
