% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_sf_trans.R
\name{flip_sf_trans}
\alias{flip_sf_trans}
\title{Flips a transect layout given as a simple 'sf' linestring.}
\usage{
flip_sf_trans(tr_sf)
}
\arguments{
\item{tr_sf}{'sf' linestring containing a transect layout.}
}
\value{
An flipped transect layout as an sf-linestring.
}
\description{
Transect layout flipped by changing the order of the nodes.
}
\details{
Note that number of nodes must be even for the flip to work.
}
\examples{
\dontrun{
  library(geo)
  tibble::tribble(~y,~x,
    67, -23,
    68, -23, 68, -22,
    67, -22, 67, -21,
    68, -21, 68, -20,
    67, -20, 67, -19,
    68, -19, 68, -18,
    67, -18, 67, -17,
    68, -17, 68, -16,
    67, -16, 67, -15,
    68, -15, 68, -14,
    67, -14) \%>\%
 select(x,y) -> l
 l <- st_linestring(as.matrix(l))
 l \%>\% st_sfc(crs=4326) -> l
 plot(l,axes=TRUE,lwd=5,col="blue",lty=2)
 plot(flip_sf_trans(l),col="magenta",lwd=3,add=TRUE)
}
}
\keyword{transect}
