% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_utils.R
\name{getBaseline}
\alias{getBaseline}
\title{Get RStox baselin}
\usage{
getBaseline(
  projectName,
  input = c("par", "proc"),
  proc = "all",
  drop = TRUE,
  startProcess = 1,
  endProcess = Inf,
  reset = FALSE,
  save = FALSE,
  modelType = "baseline",
  msg = TRUE,
  exportCSV = FALSE,
  warningLevel = 0,
  parlist = list(),
  tempRScriptFileName = NULL,
  close = FALSE,
  ...
)
}
\description{
Get RStox baselin
}
