% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_utils.R
\name{getPlots}
\alias{getPlots}
\title{Get StoX plots}
\usage{
getPlots(projectName, out = "all", options = "", ...)
}
\arguments{
\item{projectName}{Stox project name.}

\item{out}{Stox output container.}

\item{options}{Stox options}

\item{...}{Any other params to be passed on.}
}
\description{
Get StoX plots
}
