% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_utils.R
\name{getProcessDataTableAsDataFrame}
\alias{getProcessDataTableAsDataFrame}
\title{Get StoX process data table as dataframe.}
\usage{
getProcessDataTableAsDataFrame(projectName, tableName)
}
\description{
Get StoX process data table as dataframe.
}
\keyword{internal}
