% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_utils.R
\name{getProject}
\alias{getProject}
\title{Get StoX project.}
\usage{
getProject(
  projectName,
  out = c("project", "baseline", "baseline-report", "name"),
  msg = FALSE
)
}
\description{
Get StoX project.
}
