% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx2sf.R
\name{gpx2sf}
\alias{gpx2sf}
\title{Reads strata from StoX project files.}
\usage{
gpx2sf(gpx_strata_dir)
}
\arguments{
\item{gpx_strata_dir}{Path to a directory with gpx-files of stratum boundaries}
}
\value{
An sf-dataframe.
}
\description{
With some XML-magic from Arne Johannes Holmin we get at a StoX-project's
strata boundaries and read them in as an sf-data.frame.
}
\details{
It is assumed a simple route demarcating an open boundary of a polygon
is supplied, the function reads in the GPXs, closes the boundaries and
converts to 'simple features'.
}
\examples{
strata <- gpx2sf(
  paste(system.file("extdata",package="hafrostox"),"gpx",sep="/"))
strata
}
\keyword{StoX}
\keyword{XML,}
\keyword{project}
\keyword{strata,}
\keyword{stratum,}
