% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_rte.R
\name{gpx_rte}
\alias{gpx_rte}
\title{Writes a gpx-file of a single, simple route.}
\usage{
gpx_rte(file, lat, lon, name)
}
\arguments{
\item{file}{Name of gpx output file}

\item{lat, lon}{Latitude and longitude of route-points.}

\item{name}{Name of route-point, e.g 1 thru number of rte-points.}
}
\value{
Writes a gpx-file of a single route.
}
\description{
Writes a gpx-file of a single, simple route.
}
\examples{
  \dontrun{gpx_rte(tibble(
     file="tmp_rte.gpx",lat=c(65,66,67),lon=c(-20,-19,-18),name=1:3)
  } 
}
\keyword{gpx,}
\keyword{lat,}
\keyword{lon,}
\keyword{route}
