% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_wpt.R
\name{gpx_wpt}
\alias{gpx_wpt}
\title{Writes a gpx-file of waypoints with symbols from OpenCPN vocabulary.}
\usage{
gpx_wpt(file, lat, lon, name, symbol)
}
\arguments{
\item{file}{Name of gpx output file}

\item{lat, lon}{Latitude and longitude of waypoints.}

\item{name}{Name of waypoint, e.g 1 thru number of waypoints.}

\item{symbol}{Symbols e.g. from OpenCPN vocabulary (could be worked in).}
}
\description{
Writes a gpx-file of waypoints with symbols from OpenCPN vocabulary.
}
\examples{
  \dontrun{gpx_wpt(tibble(
     file="tmp_wpt.gpx",lat=c(65,66,67),lon=c(-20,-19,-18),name=1:3,
     symbol=c("Symbol-Glow-Small-Red","Symbol-Glow-Small-Blue",
   "Symbol-Glow-Small-Orange")
  } 
}
\keyword{Writes}
\keyword{a}
\keyword{gpx,}
\keyword{gpx-file}
\keyword{lat,}
\keyword{lon,}
\keyword{of}
\keyword{waypoint}
\keyword{waypoints.}
