% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_utils.R
\name{openProject}
\alias{openProject}
\title{Open StoX project.}
\usage{
openProject(
  projectName = NULL,
  out = c("project", "baseline", "baseline-report", "name"),
  msg = FALSE
)
}
\description{
Open StoX project.
}
