% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para2zz_df.R
\name{para2zz_df}
\alias{para2zz_df}
\title{Changes a parallel transect layout to a zigzag layout.}
\usage{
para2zz_df(tr_df)
}
\arguments{
\item{tr_df}{Dataframe containing a parallel transect layout.}
}
\value{
A zigzag transect layout as a dataframe.
}
\description{
Parallel layout changed to zigzag by averaging the position
of transverse transect endpoints, which become the endpoints
of the sawtooths in the zigzag.
}
\details{
The first and last nodes in the original layout are kept,
number of nodes must be even. Assuming a regular parallel layout
as input, the angle of the resulting sawtooths will be halved
at the start end end of the output zigzag.
}
\examples{
\dontrun{
  library(geo)
  tibble::tribble(~lat,~lon,
    67, -23,
    68, -23, 68, -22,
    67, -22, 67, -21,
    68, -21, 68, -20,
    67, -20, 67, -19,
    68, -19, 68, -18,
    67, -18, 67, -17,
    68, -17, 68, -16,
    67, -16, 67, -15,
    68, -15, 68, -14,
    67, -14) -> df
 geo::geoplot(df,type="n",grid=F,xlim=list(lat=c(66,68.25),lon=c(-24,-13)))
 geo::gbplot(c(500,1000),lty=2)
 geo::geolines(df,col="blue",lwd=3)
 geo::geolines(para2zz_df(df),col="magenta",lwd=3)
}
}
\keyword{parallel,}
\keyword{transect,}
\keyword{zigzag}
