% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_strata_from_project_json.R
\name{read_strata_from_project_json}
\alias{read_strata_from_project_json}
\title{Reads strata from StoX json project files.}
\usage{
read_strata_from_project_json(proj_fn)
}
\arguments{
\item{proj_fn}{Path to and filname of a StoX project json-file}
}
\value{
An sf-dataframe.
}
\description{
Uses 'rjson' to read project.json (StoX > 3) and extract
geojson strata definition, then convert with 'geojsonsf::geojson_sf'
to a simple features,'sf', object.
}
\examples{
read_strata_from_project_json(
  paste(system.file("extdata", package="hafrostox"),
    "project.json", sep="/"))
}
\keyword{StoX}
\keyword{XML,}
\keyword{project}
\keyword{strata,}
\keyword{stratum,}
