% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_strata_from_project_json_old.R
\name{read_strata_from_project_json_old}
\alias{read_strata_from_project_json_old}
\title{Reads strata from StoX project json file.}
\usage{
read_strata_from_project_json_old(project)
}
\arguments{
\item{project}{Full path to a StoX project}
}
\value{
An sf-dataframe.
read_strata_from_project_json(
paste(system.file("extdata", package="hafrostox"),
"project.json", sep="/"))
}
\description{
Simple approach of searching for strings before and after
geojson strata definition in a StoX (>3) project json file,
read directly with 'geojson_sf' into an simple features,'sf', object.
}
\keyword{StoX}
\keyword{XML,}
\keyword{project}
\keyword{strata,}
\keyword{stratum,}
