% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_utils.R
\name{runBootstrap}
\alias{runBootstrap}
\title{Run StoX bootstrap.}
\usage{
runBootstrap(
  projectName,
  bootstrapMethod = "AcousticTrawl",
  acousticMethod = PSU ~ Stratum,
  bioticMethod = PSU ~ Stratum,
  nboot = 5,
  startProcess = "TotalLengthDist",
  endProcess = "SuperIndAbundance",
  seed = 1,
  cores = 1,
  msg = TRUE,
  sorted = TRUE,
  JavaMem = getRstoxDef("JavaMem"),
  ...
)
}
\description{
Run StoX bootstrap.
}
