% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_utils.R
\name{runBootstrap_1.6}
\alias{runBootstrap_1.6}
\title{Run StoX bootstrap, version 1.6.}
\usage{
runBootstrap_1.6(
  projectName,
  bootstrapMethod = "AcousticTrawl",
  acousticMethod = PSU ~ Stratum,
  bioticMethod = PSU ~ Stratum,
  nboot = 5,
  startProcess = "TotalLengthDist",
  endProcess = "SuperIndAbundance",
  seed = 1,
  cores = 1,
  msg = TRUE,
  ...
)
}
\description{
Run StoX bootstrap, version 1.6.
}
