% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2tbl.R
\name{sf2tbl}
\alias{sf2tbl}
\title{Converts a sf geometry object to a tibble with columns 'lat' and 'lon'.}
\usage{
sf2tbl(sf)
}
\arguments{
\item{sf}{sf geometry object.}
}
\value{
A tibble with columns 'lat' and 'lon'.
}
\description{
Deals with simple/single geometries only.
}
\keyword{geometry,}
\keyword{lat,}
\keyword{lon}
\keyword{tibble,}
