% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2wkt.R
\name{sf2wkt}
\alias{sf2wkt}
\title{Writes e.g. stratum boundaries as simple feature geometries
out to well know text files.}
\usage{
sf2wkt(sfdf, fn = "tmp_wkt.txt")
}
\arguments{
\item{sfdf}{Simple features dataframe}

\item{fn}{Filename to write out to}
}
\value{
A text file representation of the geometries
}
\description{
Takes a sf-data.frame and writes out a text representation
of the geometries.
}
\examples{
\dontrun{
strata <- gpx2sf(
  paste(system.file("extdata", package="hafrostox"), "gpx", sep="/"))
sf2wkt(strata, "strata_wkt.txt")
}
}
\keyword{known}
\keyword{sf-data.frame,}
\keyword{text}
\keyword{well}
