% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv_plnr.R
\name{srv_plnr}
\alias{srv_plnr}
\title{For a given data.frame of stratum boundaries and
a set of parameters in a named list set up a survey
plan with 'Rstox::surveyPlanner'.}
\usage{
srv_plnr(stratum_df, ...)
}
\arguments{
\item{stratum_df}{Stratum boundary for a single stratum a tibble/data.frame with cols 'lat' and 'lon'}

\item{...}{Further arguments to 'Rstox::surveyPlanner'.}
}
\value{
A survey plan created with 'Rstox::surveyPlanner'.
}
\description{
For a given data.frame of stratum boundaries and
a set of parameters in a named list set up a survey
plan with 'Rstox::surveyPlanner'.
}
\examples{
\dontrun{
strata <- gpx2sf(
  paste(system.file("extdata", package="hafrostox"), "gpx", sep="/"))
sf2wkt(strata, "strata_wkt.txt")
}
}
\keyword{parameters}
\keyword{stratum,}
\keyword{surveyPlanner,}
