% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2sf.R
\name{tbl2sf}
\alias{tbl2sf}
\title{Converts a tibble with lat lon to sf geometry object.}
\usage{
tbl2sf(tbl, geom = "LINESTRING")
}
\arguments{
\item{tbl}{Tibble with columns 'lat' and 'lon'}

\item{geom}{Geometry type}
}
\value{
sf-object of requested type.
}
\description{
Returns simple geometries only, of type 'LINESTRING','MULTIPOINT' or 'POLYGON'.
}
\keyword{geometry,}
\keyword{lat,}
\keyword{lon}
\keyword{tibble,}
