% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_strata_from_project_xml.R
\name{read_strata_from_project_xml}
\alias{read_strata_from_project_xml}
\title{Reads strata from StoX project XML file.}
\usage{
read_strata_from_project_xml(project, names = NULL)
}
\arguments{
\item{project}{Full path to and filename of a StoX project}

\item{names}{Optional list of names for the individual strata}
}
\value{
An sf-dataframe.
}
\description{
With some XML-magic from Arne Johannes Holmin we get at a StoX-project's
strata boundaries and read them in as an sf-data.frame.
}
\examples{
\dontrun{
read_strata_from_project_xml(
  "~/workspace/stox/project/Test_Rstox/process/project.xml"  
}
}
\keyword{StoX}
\keyword{XML,}
\keyword{project}
\keyword{strata,}
\keyword{stratum,}
