% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans2gpx.R
\name{trans2gpx}
\alias{trans2gpx}
\title{Splits transects by strata and writes gpx-files to a given directory.}
\usage{
trans2gpx(surv, dir_name)
}
\arguments{
\item{file}{Name of gpx output file}

\item{lat, lon}{Latitude and longitude of route-points.}

\item{name}{Name of route-point, e.g 1 thru number of rte-points.}
}
\value{
Writes a gpx-file of a single route.
}
\description{
Splits transects by strata and writes gpx-files to a given directory.
}
\examples{
  strata <- gpx2sf(
    paste(system.file("extdata",package="hafrostox"),"gpx",sep="/"))
  fn <- tempfile()
  sf2wkt(strata,fn)
  zz <- surveyPlanner(fn,type="RectEnclZZ", bearing=c("NE","W","N","E"),
    nmi=1000,knots=10,seed=1,equalEffort=TRUE,retour=TRUE)
  dn <- tempdir()
  dn
  trans2gpx(zz, dn)
}
\keyword{gpx,}
\keyword{lat,}
\keyword{lon,}
\keyword{route}
